fix #64 - crash on startup
RegisterTraceGuidsW was incorrectly declared as using cdecl calling convention.
This was causing the stack to be misaligned, and then a security check failed.
Windows code will sometimes execute int 0x29 when an error is detected, this is
what was happening here.

Secondly, some code changed that used to check the return code of
GetEnvironmentVariable, but now ignores the return code and checks the value of
GetLastError. On Windows that would usually be set to ERROR_ENVVAR_NOT_FOUND if
a variable wasn't set - I never bothered setting it because nobody checked it,
but now it needs that.