Added ARCH target flag for the peloader library. Added first test which successfully check a PE (x64) header.
The peloader library can be built passing the ARCH flag as argument, which allows the user
to choose between x86 or x64. This flag is available for the peloader library only, but
the idea is to expand it to the whole project.

check_nt_hdr function has been changed in order to recognize x64 NT headers magic values.
For this purpose, a Check unit test has been added.